Avoid swift_beginAccess calls in ByteBuffer
Motivation:

After some careful auditing we noticed that access to the _Storage.bytes
field was triggering calls to swift_beginAccess on the _setBytes call
path. This isn't ideal: it's a quite heavy function, and not necessary
here as we cannot possibly violate exclusive access rules.
Unfortunately, the compiler clearly can't see that in 5.3.

(Sidebar, the compiler _can_ apparently see it in the nightly builds,
though how effective it is at it seems to be fluctuating wildly.)

With some careful work we are able to successfully elide the call by
redefining how we get at the pointer.

Modifications:

- Make the storage for `bytes` fully private.
- Provide a get-only accessor that is usableFromInline.

Result:

Fewer calls to swift_beginAccess, better performance in most benchmarks.